% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{ae_available_editions}
\alias{ae_available_editions}
\title{Return a list of available editions}
\usage{
ae_available_editions(metadata, id)
}
\arguments{
\item{id}{identifier for the dataset}
}
\description{
.. content for \details{} ..##' @title
}
\author{
Neale Swinnerton \href{mailto:neale@mastodonc.com}{neale@mastodonc.com}
}
