% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{ae_download}
\alias{ae_download}
\title{Download}
\usage{
ae_download(metadata, format = "csv")
}
\arguments{
\item{metadata}{data describing the download}

\item{format}{a valid format for the download}
}
\description{
\code{ae_download} retrieves the data described by the given df
}
